+++ src/Tests/WB.Core.SharedKernels.DataCollection.Tests/InterviewTests/when_length_of_selected_values_more_than_max_for_selected_answer_options_for_multiquestion_throw_interview_exception.cs
@@ -39,7 +39,7 @@ namespace WB.Core.SharedKernels.DataCollection.Tests.InterviewTests
         };
 
         private Because of = () => expectedException = Catch.Exception(() =>
-            interview.AnswerMultipleOptionsQuestion(userId, validatingQuestionId, new int[] { }, DateTime.Now, new decimal[] { 1, 2, 3 }));
+            interview.AnswerMultipleOptionsQuestion(userId, validatingQuestionId, new decimal[] { }, DateTime.Now, new decimal[] { 1, 2, 3 }));
 
         It should_raise_MultipleOptionsQuestionAnswered_event_with_QuestionId_equal_to_validatingQuestionId = () =>
             expectedException.ShouldBeOfType(typeof(InterviewException));
