+++ src/Tests/WB.Core.SharedKernels.DataCollection.Tests/InterviewTests/when_reevaluating_whole_interview_and_questionnaire_has_mandatory_date_and_gps_question.cs
@@ -66,7 +66,7 @@ namespace WB.Core.SharedKernels.DataCollection.Tests.InterviewTests
             eventContext = new EventContext();
 
             interview = CreateInterview(questionnaireId: questionnaireId);
-            interview.Apply(new GeoLocationQuestionAnswered(userId, geoQuestionId, new int[0], DateTime.Now, 0.111,0.222,333,new DateTimeOffset(DateTime.Now)));
+            interview.Apply(new GeoLocationQuestionAnswered(userId, geoQuestionId, new decimal[0], DateTime.Now, 0.111, 0.222, 333, new DateTimeOffset(DateTime.Now)));
         };
 
         Cleanup stuff = () =>
